DBW: fix logging sink selection, SD row commit, and RC channel mapping
Addresses the DBW software items from the 2026-08-03 status email.

Logging
- Logger: pick the CSV sink by TYPE at compile time instead of comparing
  the objects. `serialLOG == logfile` invoked operator bool() on both
  sides; for SerialUSB that reports "is a USB host attached", false at
  boot, so the SD branch was selected by mistake and logMethod fell to 2,
  silencing the CSV entirely.
- Route rows through a Print* sink so an SD failure falls back to a
  serial port instead of writing nothing anywhere ("ensure log files are
  always written").
- Restore the per-row flush dropped in e2e168f. That commit removed the
  shared serialLOG.flush() to stop SerialUSB.flush() blocking the control
  loop, and silently took SD's row commit with it, leaving a 0-byte
  LOGnn.CSV. flushSink() flushes File and compiles to nothing for serial,
  so the loop-freeze fix stays intact.
- Apply the USB host check only when the sink really is SerialUSB. The
  old `#ifdef USE_NATIVE_USB` guard gated every sink, so on a Bridge board
  the SD card only got rows while a PC had the USB port open.
- Replace the dead `#if (serialLOG != logfile)` block, which never
  compiled (the preprocessor reads both names as 0), so Serial1/2/3 never
  had begin() called.
- Write the CSV header on the first update() rather than in initialize().
  On native USB the header was emitted before the host opened the port
  and discarded, leaving rows with no column names.
- Queue boot diagnostics and emit them on the native USB port, for the
  Bridge board where the programming port is not cabled to a PC.

RTC and file naming
- initRTC() now sets the clock from build time when the DS1307 is stopped
  or unreadable; RTC.set() was commented out, so the clock was never
  corrected and every log carried a meaningless timestamp.
- Name the log MMDDHHMM.CSV from the RTC, falling back to LOGnn.CSV when
  no clock is available.

New log columns
- Operator input: op_steer (A7), op_throttle (A8) and the four switches.
- Received CAN: nav_speed, nav_brake, nav_mode, nav_angle, nav_status.
- Steering: log STEER_METHOD plus the pins that method actually drives,
  instead of always logging Rturn/Lturn, which belong to a scheme that
  was never in use.

Control fixes
- SteeringController: `#elif (STEER_METHOD == STR_MOTOR_CONTROL)` repeated
  the `#if` condition, so the two-wire L_TURN/R_TURN branch could not
  compile under any setting. Now tests SRT_HBRIDGE.
- RC_Controller: getMappedValue() rotated channels by +/-3 while
  getEtime() did not. Vehicle::updateRC() therefore read CH5 for throttle
  and CH4 (which holds the AutoMode) for steering, so the steering stick
  had no effect on desired_angle_DegX10. Verified against logged data:
  the computed value for every channel matched the rotated column exactly.
- MAP(): cast to long before subtracting. Callers pass unsigned long
  elapsedTime[], so a dead channel reading 0 us wrapped to 4294867
  instead of going negative. Dead channels now map to 0.
- Drive_By_Wire.ino: always begin Serial. Under USE_NATIVE_USB it was
  never begun, yet initRTC()/initSD()/openSD() print to it, and
  UARTClass::write() spin-locks once its buffer fills with no drain
  interrupt enabled.

Verified: 15 build combinations (3 steering methods x 4 sinks) compile
clean. Column parity and the logging changes confirmed on hardware; the
SD flush, dated filename and RTC set paths are compile-verified only,
pending a board with a working card and DS1307.

+++ Drive_By_Wire/DBW_Pins.h
@@ -12,7 +12,10 @@
 //Version can be 1,3, 4, 5; LLB = Low Level Board = Drive By Wire; 1,2,3 no longer supported.
 #define DBWversion 5
 
-#define MAP(val,dlo,dhi,nlo,nhi) ((nlo) + (((val-(dlo))*((nhi)-(nlo)))/((dhi)-(dlo))))
+// Cast to long before subtracting: callers pass unsigned long (RC elapsedTime[]),
+// and when val < dlo the unsigned subtraction wraps to ~4.29e9 instead of going
+// negative. A dead RC channel reading 0 us used to map to 4294867 this way.
+#define MAP(val,dlo,dhi,nlo,nhi) ((nlo) + ((((long)(val)-(long)(dlo))*((nhi)-(nlo)))/((dhi)-(dlo))))
 
 // serial goes to serial monitor; other settings can have another microprocessor do the logging
 // serial1 connects to router on Bridge;
